requirements.txt
aiohttp==3.7.4.post0
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
braceexpand==0.1.7
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cycler==0.10.0
discord==1.7.3
discord-ext-bot==1.0.1
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@fbb8803779373357e274e1540b368365fd9d8074
discord-together @ git+https://github.com/walthersmith/discord-together@7a5b5a8bce8f2fe8563f6ae96e8c7484cb04bd42
discord.py==1.7.3
emoji==1.4.2
googletrans==3.0.0
greenlet==1.1.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
import-expression==1.1.4
jishaku==2.2.0
kiwisolver==1.3.1
matplotlib==3.4.2
multidict==5.1.0
numpy==1.21.1
pandas==1.3.1
parsedatetime==2.6
Pillow==8.3.1
psutil==5.8.0
pycparser==2.20
pygit2==1.6.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.22
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
